Add RedisPublisherBolt and logging to URLCounterBolt
