+++ RTwUP/src/main/java/storm/RtwupTopology.java
@@ -4,6 +4,7 @@
 package storm;
 
 import storm.bolts.ExpanderBolt; 
+import storm.bolts.RedisPublisherBolt;
 import storm.bolts.URLCounterBolt;
 import storm.spouts.TwitterSpout;
 import backtype.storm.Config;
@@ -29,6 +30,9 @@ public class RtwupTopology {
 				"filteredStream");
 		builder.setBolt("urlCounter", new URLCounterBolt(), 5).fieldsGrouping(
 				"expander", new Fields("expanded_url_domain"));
+		builder.setBolt("RedisPublisher", new RedisPublisherBolt(), 1).shuffleGrouping(
+				"urlCounter");
+
 		
 		Config conf = new Config();
 		conf.setDebug(true);
