+++ RTwUP/src/main/java/storm/bolts/RedisPublisherBolt.java
@@ -1,0 +1,52 @@
+package storm.bolts;
+
+import java.util.Map;
+
+import redis.clients.jedis.Jedis;
+import redis.clients.jedis.JedisPool;
+import redis.clients.jedis.JedisPoolConfig;
+import view.PageDictionary;
+import backtype.storm.task.TopologyContext;
+import backtype.storm.topology.BasicOutputCollector;
+import backtype.storm.topology.OutputFieldsDeclarer;
+import backtype.storm.topology.base.BaseBasicBolt;
+import backtype.storm.tuple.Tuple;
+
+/**
+ * This bolt publishes the URL ranking to Redis.
+ * 
+ * @author Gabriele de Capoa, Gabriele Proni
+ *
+ */
+
+public class RedisPublisherBolt extends BaseBasicBolt{
+
+	private static final long serialVersionUID = 1L;
+	private JedisPool pool = null;
+	private Jedis jedis = null;
+	
+	@Override
+	public void prepare(Map conf, TopologyContext context){
+		this.pool = new JedisPool(new JedisPoolConfig(), "localhost");
+		this.jedis = this.pool.getResource();
+	}
+	
+	@Override
+	public void execute(Tuple input, BasicOutputCollector collector) {
+				
+		String ranking = PageDictionary.getInstance(7).getTopNelementsStringified();
+		
+		this.jedis.publish("RTwUP", ranking);
+	}
+
+	@Override
+	public void declareOutputFields(OutputFieldsDeclarer declarer) {
+	}
+
+	@Override
+	public void cleanup(){
+		this.pool.returnResource(this.jedis);
+		this.pool.destroy();
+	}
+	
+}
